+++ previous-versions/accounting/pkg/models/operations/getcreateupdatebankaccountsmodel.go
@@ -8,7 +8,9 @@ import (
 )
 
 type GetCreateUpdateBankAccountsModelRequest struct {
-	CompanyID    string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a company.
+	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a connection.
 	ConnectionID string `pathParam:"style=simple,explode=false,name=connectionId"`
 }
 
@@ -27,12 +29,15 @@ func (o *GetCreateUpdateBankAccountsModelRequest) GetConnectionID() string {
 }
 
 type GetCreateUpdateBankAccountsModelResponse struct {
+	// HTTP response content type for this operation
 	ContentType string
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
 	// OK
-	PushOption  *shared.PushOption
-	StatusCode  int
+	PushOption *shared.PushOption
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
 	RawResponse *http.Response
 }