+++ previous-versions/accounting/pkg/models/operations/getdirectincome.go
@@ -8,8 +8,11 @@ import (
 )
 
 type GetDirectIncomeRequest struct {
-	CompanyID      string `pathParam:"style=simple,explode=false,name=companyId"`
-	ConnectionID   string `pathParam:"style=simple,explode=false,name=connectionId"`
+	// Unique identifier for a company.
+	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a connection.
+	ConnectionID string `pathParam:"style=simple,explode=false,name=connectionId"`
+	// Unique identifier for a direct income.
 	DirectIncomeID string `pathParam:"style=simple,explode=false,name=directIncomeId"`
 }
 
@@ -35,13 +38,16 @@ func (o *GetDirectIncomeRequest) GetDirectIncomeID() string {
 }
 
 type GetDirectIncomeResponse struct {
+	// HTTP response content type for this operation
 	ContentType string
 	// Success
 	DirectIncome *shared.DirectIncome
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
-	StatusCode   int
-	RawResponse  *http.Response
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
+	RawResponse *http.Response
 }
 
 func (o *GetDirectIncomeResponse) GetContentType() string {