+++ previous-versions/accounting/pkg/models/operations/getsupplierattachment.go
@@ -8,11 +8,13 @@ import (
 )
 
 type GetSupplierAttachmentRequest struct {
-	// Unique identifier for an attachment
+	// Unique identifier for an attachment.
 	AttachmentID string `pathParam:"style=simple,explode=false,name=attachmentId"`
-	CompanyID    string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a company.
+	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a connection.
 	ConnectionID string `pathParam:"style=simple,explode=false,name=connectionId"`
-	// Unique identifier for a supplier
+	// Unique identifier for a supplier.
 	SupplierID string `pathParam:"style=simple,explode=false,name=supplierId"`
 }
 
@@ -46,12 +48,15 @@ func (o *GetSupplierAttachmentRequest) GetSupplierID() string {
 
 type GetSupplierAttachmentResponse struct {
 	// Success
-	Attachment  *shared.Attachment
+	Attachment *shared.Attachment
+	// HTTP response content type for this operation
 	ContentType string
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
-	StatusCode   int
-	RawResponse  *http.Response
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
+	RawResponse *http.Response
 }
 
 func (o *GetSupplierAttachmentResponse) GetAttachment() *shared.Attachment {