+++ previous-versions/accounting/pkg/models/operations/uploaddirectcostattachment.go
@@ -27,10 +27,12 @@ func (o *UploadDirectCostAttachmentRequestBody) GetRequestBody() string {
 }
 
 type UploadDirectCostAttachmentRequest struct {
-	RequestBody  *UploadDirectCostAttachmentRequestBody `multipartForm:"file" request:"mediaType=multipart/form-data"`
-	CompanyID    string                                 `pathParam:"style=simple,explode=false,name=companyId"`
-	ConnectionID string                                 `pathParam:"style=simple,explode=false,name=connectionId"`
-	// Unique identifier for a direct cost
+	RequestBody *UploadDirectCostAttachmentRequestBody `multipartForm:"file" request:"mediaType=multipart/form-data"`
+	// Unique identifier for a company.
+	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a connection.
+	ConnectionID string `pathParam:"style=simple,explode=false,name=connectionId"`
+	// Unique identifier for a direct cost.
 	DirectCostID string `pathParam:"style=simple,explode=false,name=directCostId"`
 }
 
@@ -63,11 +65,14 @@ func (o *UploadDirectCostAttachmentRequest) GetDirectCostID() string {
 }
 
 type UploadDirectCostAttachmentResponse struct {
+	// HTTP response content type for this operation
 	ContentType string
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
-	StatusCode   int
-	RawResponse  *http.Response
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
+	RawResponse *http.Response
 }
 
 func (o *UploadDirectCostAttachmentResponse) GetContentType() string {