+++ previous-versions/accounting/pkg/models/shared/account.go
@@ -3,7 +3,8 @@
 package shared
 
 import (
-	"github.com/codatio/client-sdk-go/previous-versions/accounting/pkg/types"
+	"github.com/codatio/client-sdk-go/previous-versions/accounting/pkg/utils"
+	"github.com/ericlagergren/decimal"
 )
 
 // Account - > **Language tip:** Accounts are also referred to as **chart of accounts**, **nominal accounts**, and **general ledger**.
@@ -44,7 +45,7 @@ type Account struct {
 	// There are only a very small number of edge cases where this currency code is returned by the Codat system.
 	Currency *string `json:"currency,omitempty"`
 	// Current balance in the account.
-	CurrentBalance *types.Decimal `json:"currentBalance,omitempty"`
+	CurrentBalance *decimal.Big `decimal:"number" json:"currentBalance,omitempty"`
 	// Description for the account.
 	Description *string `json:"description,omitempty"`
 	// Full category of the account.
@@ -75,6 +76,17 @@ type Account struct {
 	ValidDatatypeLinks []ValidDataTypeLinks `json:"validDatatypeLinks,omitempty"`
 }
 
+func (a Account) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(a, "", false)
+}
+
+func (a *Account) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &a, "", false, false); err != nil {
+		return err
+	}
+	return nil
+}
+
 func (o *Account) GetCurrency() *string {
 	if o == nil {
 		return nil
@@ -82,7 +94,7 @@ func (o *Account) GetCurrency() *string {
 	return o.Currency
 }
 
-func (o *Account) GetCurrentBalance() *types.Decimal {
+func (o *Account) GetCurrentBalance() *decimal.Big {
 	if o == nil {
 		return nil
 	}