+++ previous-versions/accounting/pkg/models/shared/billcreditnotelineitem.go
@@ -3,7 +3,7 @@
 package shared
 
 import (
-	"github.com/codatio/client-sdk-go/previous-versions/accounting/pkg/types"
+	"github.com/codatio/client-sdk-go/previous-versions/accounting/pkg/utils"
 	"github.com/ericlagergren/decimal"
 )
 
@@ -11,9 +11,11 @@ import (
 type BillCreditNoteLineItemTracking struct {
 	CategoryRefs []TrackingCategoryRef  `json:"categoryRefs"`
 	CustomerRef  *AccountingCustomerRef `json:"customerRef,omitempty"`
-	IsBilledTo   BilledToType           `json:"isBilledTo"`
-	IsRebilledTo BilledToType           `json:"isRebilledTo"`
-	ProjectRef   *ProjectRef            `json:"projectRef,omitempty"`
+	// Defines if the invoice or credit note is billed/rebilled to a project or customer.
+	IsBilledTo BilledToType `json:"isBilledTo"`
+	// Defines if the invoice or credit note is billed/rebilled to a project or customer.
+	IsRebilledTo BilledToType `json:"isRebilledTo"`
+	ProjectRef   *ProjectRef  `json:"projectRef,omitempty"`
 }
 
 func (o *BillCreditNoteLineItemTracking) GetCategoryRefs() []TrackingCategoryRef {
@@ -57,16 +59,16 @@ type BillCreditNoteLineItem struct {
 	// Friendly name of each line item. For example, the goods or service for which credit has been received.
 	Description *string `json:"description,omitempty"`
 	// Value of any discounts applied.
-	DiscountAmount *types.Decimal `json:"discountAmount,omitempty"`
+	DiscountAmount *decimal.Big `decimal:"number" json:"discountAmount,omitempty"`
 	// Percentage rate of any discount applied to the line item.
-	DiscountPercentage *types.Decimal `json:"discountPercentage,omitempty"`
-	ItemRef            *ItemRef       `json:"itemRef,omitempty"`
+	DiscountPercentage *decimal.Big `decimal:"number" json:"discountPercentage,omitempty"`
+	ItemRef            *ItemRef     `json:"itemRef,omitempty"`
 	// Number of units of the goods or service for which credit has been received.
-	Quantity types.Decimal `json:"quantity"`
+	Quantity *decimal.Big `decimal:"number" json:"quantity"`
 	// Amount of credit associated with the line item, including discounts but excluding tax.
-	SubTotal *types.Decimal `json:"subTotal,omitempty"`
+	SubTotal *decimal.Big `decimal:"number" json:"subTotal,omitempty"`
 	// Amount of tax associated with the line item.
-	TaxAmount *types.Decimal `json:"taxAmount,omitempty"`
+	TaxAmount *decimal.Big `decimal:"number" json:"taxAmount,omitempty"`
 	// Data types that reference a tax rate, for example invoice and bill line items, use a taxRateRef that includes the ID and name of the linked tax rate.
 	//
 	// Found on:
@@ -79,7 +81,7 @@ type BillCreditNoteLineItem struct {
 	// - Items
 	TaxRateRef *TaxRateRef `json:"taxRateRef,omitempty"`
 	// Total amount of the line item, including discounts and tax.
-	TotalAmount *types.Decimal `json:"totalAmount,omitempty"`
+	TotalAmount *decimal.Big `decimal:"number" json:"totalAmount,omitempty"`
 	// Categories, and a project and customer, against which the item is tracked.
 	Tracking *BillCreditNoteLineItemTracking `json:"tracking,omitempty"`
 	// Reference to the tracking categories to which the line item is linked.
@@ -87,7 +89,18 @@ type BillCreditNoteLineItem struct {
 	// Deprecated field: This will be removed in a future release, please migrate away from it as soon as possible.
 	TrackingCategoryRefs []TrackingCategoryRef `json:"trackingCategoryRefs,omitempty"`
 	// Unit price of the goods or service.
-	UnitAmount types.Decimal `json:"unitAmount"`
+	UnitAmount *decimal.Big `decimal:"number" json:"unitAmount"`
+}
+
+func (b BillCreditNoteLineItem) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(b, "", false)
+}
+
+func (b *BillCreditNoteLineItem) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &b, "", false, false); err != nil {
+		return err
+	}
+	return nil
 }
 
 func (o *BillCreditNoteLineItem) GetAccountRef() *AccountRef {
@@ -104,14 +117,14 @@ func (o *BillCreditNoteLineItem) GetDescription() *string {
 	return o.Description
 }
 
-func (o *BillCreditNoteLineItem) GetDiscountAmount() *types.Decimal {
+func (o *BillCreditNoteLineItem) GetDiscountAmount() *decimal.Big {
 	if o == nil {
 		return nil
 	}
 	return o.DiscountAmount
 }
 
-func (o *BillCreditNoteLineItem) GetDiscountPercentage() *types.Decimal {
+func (o *BillCreditNoteLineItem) GetDiscountPercentage() *decimal.Big {
 	if o == nil {
 		return nil
 	}
@@ -125,21 +138,21 @@ func (o *BillCreditNoteLineItem) GetItemRef() *ItemRef {
 	return o.ItemRef
 }
 
-func (o *BillCreditNoteLineItem) GetQuantity() types.Decimal {
+func (o *BillCreditNoteLineItem) GetQuantity() *decimal.Big {
 	if o == nil {
-		return types.Decimal{Big: *(new(decimal.Big).SetFloat64(0.0))}
+		return new(decimal.Big).SetFloat64(0.0)
 	}
 	return o.Quantity
 }
 
-func (o *BillCreditNoteLineItem) GetSubTotal() *types.Decimal {
+func (o *BillCreditNoteLineItem) GetSubTotal() *decimal.Big {
 	if o == nil {
 		return nil
 	}
 	return o.SubTotal
 }
 
-func (o *BillCreditNoteLineItem) GetTaxAmount() *types.Decimal {
+func (o *BillCreditNoteLineItem) GetTaxAmount() *decimal.Big {
 	if o == nil {
 		return nil
 	}
@@ -153,7 +166,7 @@ func (o *BillCreditNoteLineItem) GetTaxRateRef() *TaxRateRef {
 	return o.TaxRateRef
 }
 
-func (o *BillCreditNoteLineItem) GetTotalAmount() *types.Decimal {
+func (o *BillCreditNoteLineItem) GetTotalAmount() *decimal.Big {
 	if o == nil {
 		return nil
 	}
@@ -174,9 +187,9 @@ func (o *BillCreditNoteLineItem) GetTrackingCategoryRefs() []TrackingCategoryRef
 	return o.TrackingCategoryRefs
 }
 
-func (o *BillCreditNoteLineItem) GetUnitAmount() types.Decimal {
+func (o *BillCreditNoteLineItem) GetUnitAmount() *decimal.Big {
 	if o == nil {
-		return types.Decimal{Big: *(new(decimal.Big).SetFloat64(0.0))}
+		return new(decimal.Big).SetFloat64(0.0)
 	}
 	return o.UnitAmount
 }