+++ previous-versions/accounting/pkg/models/shared/items1.go
@@ -3,11 +3,14 @@
 package shared
 
 type Items1 struct {
-	Links        Links  `json:"_links"`
-	PageNumber   int64  `json:"pageNumber"`
-	PageSize     int64  `json:"pageSize"`
-	Results      []Item `json:"results,omitempty"`
-	TotalResults int64  `json:"totalResults"`
+	Links Links `json:"_links"`
+	// Current page number.
+	PageNumber int64 `json:"pageNumber"`
+	// Number of items to return in results array.
+	PageSize int64  `json:"pageSize"`
+	Results  []Item `json:"results,omitempty"`
+	// Total number of items.
+	TotalResults int64 `json:"totalResults"`
 }
 
 func (o *Items1) GetLinks() Links {