+++ previous-versions/accounting/RELEASES.md
@@ -371,3 +371,13 @@ Based on:
 - [go v0.26.0] previous-versions/accounting
 ### Releases
 - [Go v0.26.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/accounting/v0.26.0 - previous-versions/accounting
+
+## 2023-10-05 00:21:16
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.94.0 (2.147.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.27.0] previous-versions/accounting
+### Releases
+- [Go v0.27.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/accounting/v0.27.0 - previous-versions/accounting